优化ViewModel创建，优化界面互斥功能

+++ Assets/BridgeUI/Core/Graph/Editor/BridgeUIGraphCtrl.cs
@@ -125,7 +125,7 @@ namespace BridgeUI.Drawer
                 bridge.viewModel = connection.viewModel;
                 bridge.showModel = connection.show;
                 if(innode != null && innode.OutputPoints != null){
-                    bridge.index = innode.OutputPoints.FindIndex(x => x.Id == item.FromNodeConnectionPointId);
+                    bridge.index = (short)innode.OutputPoints.FindIndex(x => x.Id == item.FromNodeConnectionPointId);
                 }
                 bridges.Add(bridge);
             }
